// frontend/src/types/upscaler.ts
export enum UpscaleMode {
    CREATIVE = 'creative',
    PRECISION = 'precision',
}

export enum UpscaleModel {
    MAGNIFIC = 'magnific',
    REAL_ESRGAN = 'real-esrgan',
    GFPGAN = 'gfpgan',
    CLARITY_AI = 'clarity-ai',
    ULTRA_SHARP = 'ultra-sharp',
}

export enum UpscaleScale {
    X2 = 2,
    X4 = 4,
    X8 = 8,
    X16 = 16,
}

export enum UpscaleOptimization {
    STANDARD_ULTRA = 'standard-ultra',
    LANDSCAPE = 'landscape',
    PORTRAIT = 'portrait',
    ART_ILLUSTRATION = 'art-illustration',
    PIXEL_ART = 'pixel-art',
    CINEMATIC = 'cinematic',
}

export enum UpscaleEngine {
    AUTOMATIC = 'automatic',
    ENGINE_V1 = 'v1',
    ENGINE_V2 = 'v2',
}

export interface UpscaleParams {
    imageUrl?: string;
    mode: UpscaleMode;
    model: UpscaleModel;
    scale: UpscaleScale;
    optimization: UpscaleOptimization;
    creativity: number;
    hdr: number;
    resemblance: number;
    fractality: number;
    engine: UpscaleEngine;
    prompt?: string;
}